pyproject.toml
[tool.poetry]
name = "p10-backend"
version = "0.1.0"
description = "F1 'Guess the P10 and first DNF' game backend"
authors = ["Alejandro Brunacci <abrunacci87@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.0"
uvicorn = "^0.23.2"
pytest-cov = "^4.1.0"
httpx = "^0.25.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
